pyproject.toml
[tool.poetry]
name = "sloth-auth"
version = "0.1.0"
description = ""
authors = ["lana_remenyuk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
asyncpg = "^0.29.0"
uvicorn = "^0.32.0"
aiosmtplib = "^3.0.2"
jinja2 = "^3.1.4"
redis = "^5.1.1"
pydantic-settings = "^2.5.2"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
